fix get_dict
