+++ app/database/model/__init__.py
@@ -3,7 +3,6 @@ from .common import (
     db,
     filetimeformat,
     finishedStatus_id,
-    get_dict,
     timeformat
 )
 from .Buildings import Buildings
@@ -21,3 +20,25 @@ idTables = {t for t in allTables if "id" in t.__mapper__.columns}
 sequenceTables = {t for t in allTables if "sequence" in t.__mapper__.columns}
 
 # __all__ = []
+
+
+def get_dict(row):
+    """
+    Get pure dict from table without relationship.
+    datetime.datetime will be converted to str
+    """
+    import datetime
+
+    def process_value(value):
+        if isinstance(value, datetime.datetime):
+            return value.strftime(timeformat)
+        else:
+            return value
+
+    if type(row) not in allTables:
+        return {}
+    else:
+        return {
+            key: process_value(row.__dict__[key])
+            for key in type(row).__mapper__.columns.keys()
+        }
